The Extruct package now supports <Date> and <Datetime> types.
The following rules apply:

ISO 8601 UTC is used for the string format of a datetime:
For example: 2010-03-22T15:22:21-00:00

If Extruct <Date> encounters a string, it must begin with the "yyyy-mm-dd" format.
If Extruct <Date> encounters a datetime object, it will convert it to a date
If Extruct <Date> encounters a date object, it will pass through unharmed

If Extruct <DateTime> encounters a string, it must begin with the "yyyy-mm-ddThh:ii:ss+00:00" format.
If Extruct <DateTime> encounters a date object, it will convert it to a datetime with 0 for hours, minutes, seconds, and milliseconds
If Extruct <DateTime> encounters a datetime object, it will pass through unharmed